+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added the data gathering status column to the historical market data table of the admin control
+- Deactivated asset profiles that has been delisted by Yahoo Finance
 
 ### Changed